ImageBooru touchups
My changes to this function got a little unwieldly, so allow me to explain:

First off behavioral changes are at a minimum,
The command now returns a link to the post of the image (in bold, so it's not previewed)
The command now loops forever instead of failing (this is only bad if they turn off the api, maybe we should provide an exit for this case)

Aside from that, it's all just a bit of refactoring:
There were a lot of strings repeated, so I cleaned them up.
lolibooru has both index.xml and post.xml, since no one else had index.xml, I moved it to post.xml

I noticed the pattern of setting the two(now three) variables all in the same place, so I created a Tuple for them.
Named the tuple ImageBoard, and named its items for ease of use.

The GetBooruImageLink and GetBooruPostCount functions were way too similar,
I merged their commonality and called it GetBooruCommon, they call that now.