Add ability to ignore roles `ignore role <role name1>,<rolename2>,<rolename3>`
This command is kinda raw right now, in the future there will hopefully be a way to mention roles, and we'll stop using ? instead of @ or # to show them.

In the future, when mentioning roles is a thing this command will be combined into normal ignore command,
the existing code for this is already present, commented out.

Right now, senpai's roles can't be ignored,
While this isn't necessarily ignoring senpai, it is ignoring senpai on the server,
we may want to remove this check down the line, if nekobot becomes more of an administration tool for servers out of our command, but really, it shouldn't matter.